feat: Move resize into helper file
